backend/requirements.txt
flask==2.0.3
torch==2.0.1
torchvision==0.15.2
pymilvus==2.3.1
pymysql==1.0.3
mysql-connector-python==8.0.33
Pillow==9.5.0
transformers==4.29.2
numpy==1.24.3
packaging~=21.3
tqdm~=4.64.1
ftfy~=6.3.1
regex~=2022.7.9
pytest~=7.1.2
pandas~=1.4.4
setuptools~=63.4.1
scikit-image~=0.19.2
ipython~=7.31.1
matplotlib~=3.5.2
requests~=2.28.1
werkzeug==2.0.3
urllib3~=1.26.20